refactor(video call): extra helper functions
