+++ src/components/E2EEncryptionSupportBanner/E2EEncryptionSupportBanner.tsx
@@ -2,18 +2,13 @@ import { useState, useContext, useEffect } from 'react';
 import * as React from 'react';
 import { Banner } from '../banner/Banner';
 import './E2EEncryptionSupportBanner.styles.scss';
-import { supportsE2EEncryptionVideoCall } from '../../utils/videoCallHelpers';
-import { useTranslation } from 'react-i18next';
 import {
-	AUTHORITIES,
-	ConsultingTypeBasicInterface,
-	ConsultingTypesContext,
-	hasUserAuthority,
-	UserDataContext,
-	UserDataInterface
-} from '../../globalState';
+	hasVideoCallAbility,
+	supportsE2EEncryptionVideoCall
+} from '../../utils/videoCallHelpers';
+import { useTranslation } from 'react-i18next';
+import { ConsultingTypesContext, UserDataContext } from '../../globalState';
 import { Link } from 'react-router-dom';
-import { hasVideoCallFeature } from '../app/RouterConfig';
 
 export const E2EEncryptionSupportBanner = () => {
 	const [showBanner, setShowBanner] = useState<boolean>(
@@ -24,36 +19,8 @@ export const E2EEncryptionSupportBanner = () => {
 	const { consultingTypes } = useContext(ConsultingTypesContext);
 	const { userData } = useContext(UserDataContext);
 
-	const shouldHideBanner = (
-		userData: UserDataInterface,
-		consultingTypes: ConsultingTypeBasicInterface[]
-	) => {
-		// check if User can be called by any of his registered agencies
-		if (hasUserAuthority(AUTHORITIES.ASKER_DEFAULT, userData)) {
-			const registeredConsultingTypes = Object.values(
-				userData.consultingTypes
-			)
-				.filter((el) => el.isRegistered)
-				.map((el) => el.agency.consultingType);
-			console.log(registeredConsultingTypes, userData);
-			const userCanBeCalled = registeredConsultingTypes.some((el) =>
-				Object.values(consultingTypes).some(
-					(consultingType) =>
-						consultingType.id === el &&
-						consultingType.isVideoCallAllowed
-				)
-			);
-			if (!userCanBeCalled) {
-				return true;
-			}
-		} else if (!hasVideoCallFeature(userData, consultingTypes)) {
-			return true;
-		}
-		return false;
-	};
-
 	useEffect(() => {
-		if (shouldHideBanner(userData, consultingTypes)) {
+		if (hasVideoCallAbility(userData, consultingTypes)) {
 			setShowBanner(false);
 		}
 	}, [userData, consultingTypes]);
