+++ src/components/app/RouterConfig.tsx
@@ -45,6 +45,7 @@ import { Booking } from '../../containers/bookings/components/Booking/booking';
 import { BookingCancellation } from '../../containers/bookings/components/BookingCancellation/bookingCancellation';
 import { BookingEvents } from '../../containers/bookings/components/BookingEvents/bookingEvents';
 import { BookingReschedule } from '../../containers/bookings/components/BookingReschedule/bookingReschedule';
+import { hasVideoCallFeature } from '../../utils/videoCallHelpers';
 
 const SessionView = lazy(() =>
 	import('../session/SessionView').then((m) => ({ default: m.SessionView }))
@@ -52,18 +53,6 @@ const SessionView = lazy(() =>
 const WriteEnquiry = lazy(() =>
 	import('../enquiry/WriteEnquiry').then((m) => ({ default: m.WriteEnquiry }))
 );
-
-export const hasVideoCallFeature = (userData, consultingTypes) =>
-	userData &&
-	hasUserAuthority(AUTHORITIES.CONSULTANT_DEFAULT, userData) &&
-	userData.agencies.some(
-		(agency) =>
-			!!(consultingTypes || []).find(
-				(consultingType) =>
-					consultingType.id === agency.consultingType &&
-					consultingType.isVideoCallAllowed
-			)
-	);
 
 const showAppointmentsMenuItem = (userData, hasAssignedConsultant) => {
 	return (
