+++ src/components/notifications/Notifications.tsx
@@ -2,10 +2,9 @@ import * as React from 'react';
 import { isNotificationTypeCall } from '../incomingVideoCall/IncomingVideoCall';
 import './notifications.styles';
 import incomingCallRingtone from '../../resources/audio/incomingCall.mp3';
-
 import { NotificationType } from '../../globalState';
 import { Notification } from './Notification';
-import { E2EEncryptionSupportHelp } from '../E2EEncryptionSupportHelp/E2EEncryptionSupportHelp';
+import { supportsE2EEncryptionVideoCall } from '../../utils/videoCallHelpers';
 
 type NotificationsProps = {
 	notifications: NotificationType[];
@@ -26,7 +25,7 @@ export const Notifications = (props: NotificationsProps) => {
 					/>
 				)
 			)}
-			{hasIncomingVideoCall && !E2EEncryptionSupportHelp() && (
+			{hasIncomingVideoCall && supportsE2EEncryptionVideoCall() && (
 				<audio loop autoPlay data-cy="incoming-video-call-audio">
 					<source
 						src={
