+++ src/utils/videoCallHelpers.ts
@@ -1,4 +1,10 @@
 import { getValueFromCookie } from '../components/sessionCookie/accessSessionCookie';
+import {
+	hasUserAuthority,
+	AUTHORITIES,
+	ConsultingTypeBasicInterface,
+	UserDataInterface
+} from '../globalState';
 import { appConfig } from './appConfig';
 
 export const currentUserWasVideoCallInitiator = (initiatorRcUserId: string) =>
@@ -75,4 +81,43 @@ const supportsInsertableStreams = () => {
 	} catch {
 		return false;
 	}
+};
+
+export const hasVideoCallFeature = (userData, consultingTypes) =>
+	userData &&
+	hasUserAuthority(AUTHORITIES.CONSULTANT_DEFAULT, userData) &&
+	userData.agencies.some(
+		(agency) =>
+			!!(consultingTypes || []).find(
+				(consultingType) =>
+					consultingType.id === agency.consultingType &&
+					consultingType.isVideoCallAllowed
+			)
+	);
+
+export const hasVideoCallAbility = (
+	userData: UserDataInterface,
+	consultingTypes: ConsultingTypeBasicInterface[]
+) => {
+	// check if User can be called by any of his registered agencies
+	if (hasUserAuthority(AUTHORITIES.ASKER_DEFAULT, userData)) {
+		const registeredConsultingTypes = Object.values(
+			userData.consultingTypes
+		)
+			.filter((el) => el.isRegistered)
+			.map((el) => el.agency.consultingType);
+		const userCanBeCalled = registeredConsultingTypes.some((el) =>
+			Object.values(consultingTypes).some(
+				(consultingType) =>
+					consultingType.id === el &&
+					consultingType.isVideoCallAllowed
+			)
+		);
+		if (!userCanBeCalled) {
+			return true;
+		}
+	} else if (!hasVideoCallFeature(userData, consultingTypes)) {
+		return true;
+	}
+	return false;
 };
